# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: provider字段不能为空
        type: string
      error:
        example: true
        type: boolean
      message:
        example: 参数验证失败
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.WeChatNotificationHistoryEntry:
    properties:
      duration:
        description: 执行耗时
        example: 2m6s
        type: string
      end_time:
        description: 结束时间
        example: "2025-09-12T08:15:30Z"
        type: string
      error:
        description: 错误信息
        example: 连接超时
        type: string
      id:
        description: 任务ID
        example: task_123456
        type: string
      message:
        description: 执行消息
        example: 通知发送成功
        type: string
      records_processed:
        description: 处理记录数
        example: 1500
        type: integer
      start_time:
        description: 开始时间
        example: "2025-09-12T08:13:24Z"
        type: string
      status:
        description: 执行状态
        example: completed
        type: string
    type: object
  models.WeChatNotificationRequest:
    description: 企业微信费用报告通知触发请求，支持指定分析日期、云服务商和告警阈值
    properties:
      alert_threshold:
        description: 费用变化告警阈值（百分比），当费用变化超过此值时触发告警，默认使用系统配置值，范围：0-100
        example: 10
        maximum: 100
        minimum: 0
        type: number
      date:
        description: 分析日期，格式：YYYY-MM-DD，默认为当前日期。示例：2025-09-12
        example: "2025-09-12"
        type: string
      force_notify:
        description: 强制发送通知，设为true时即使未达到告警阈值也会发送通知，默认为false
        example: false
        type: boolean
      providers:
        description: 要分析的云服务商列表，支持：volcengine（火山云）、alicloud（阿里云）、aws、azure、gcp，默认为
          ["volcengine", "alicloud"]
        example:
        - volcengine
        - alicloud
        items:
          type: string
        type: array
      test_mode:
        description: 测试模式，设为true时仅测试企业微信连接而不发送实际的费用报告，默认为false
        example: false
        type: boolean
    type: object
  models.WeChatNotificationStatusResponse:
    properties:
      alert_threshold:
        description: 告警阈值
        example: 10
        type: number
      enabled:
        description: 微信通知功能是否启用
        example: true
        type: boolean
      executor_enabled:
        description: 执行器是否启用
        example: true
        type: boolean
      executor_status:
        description: 执行器状态
        example: ready
        type: string
      history_count:
        description: 历史记录数量
        example: 5
        type: integer
      max_retries:
        description: 最大重试次数
        example: 3
        type: integer
      mention_users:
        description: 提及用户列表
        example:
        - user1
        - user2
        items:
          type: string
        type: array
      recent_history:
        description: 最近通知历史
        items:
          $ref: '#/definitions/models.WeChatNotificationHistoryEntry'
        type: array
      retry_delay:
        description: 重试延迟时间（秒）
        example: 5
        type: integer
      timestamp:
        description: 状态查询时间戳
        example: "2025-09-12T08:13:24Z"
        type: string
      webhook_configured:
        description: Webhook是否已配置
        example: true
        type: boolean
      webhook_url:
        description: 脱敏后的Webhook URL
        example: https://qyapi***webhook
        type: string
    type: object
  models.WeChatTestRequest:
    description: 企业微信Webhook连接测试请求，用于验证Webhook URL的可用性
    properties:
      custom_message:
        description: 自定义测试消息内容，最大长度500字符，默认使用系统预设的测试消息
        example: 这是一条来自费用监控系统的测试消息
        maxLength: 500
        type: string
      timeout:
        description: 连接超时时间（秒），范围1-60秒，默认为10秒
        example: 10
        maximum: 60
        minimum: 1
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@goscan.com
    name: API Support
  description: |-
    云账单数据同步服务API文档
    支持火山云、阿里云等多云平台的账单数据自动同步到ClickHouse数据库
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Goscan API
  version: "1.0"
paths:
...
  /notifications/wechat:
    post:
      consumes:
      - application/json
      description: |-
        Manually trigger WeChat Enterprise cost report notification, supports specifying analysis date, cloud provider list and alert threshold. Can be used for temporary notification needs outside of scheduled tasks.

        **请求示例：**
        ```json
        {
        "date": "2025-09-12",
        "providers": ["volcengine", "alicloud"],
        "alert_threshold": 15.0,
        "force_notify": false,
        "test_mode": false
        }
        ```

        **Field Descriptions:**
        - `date`: Analysis date, format YYYY-MM-DD, uses current date if not provided
        - `providers`: Cloud provider list, supports volcengine/alicloud/aws/azure/gcp
        - `alert_threshold`: Cost change alert threshold (percentage), range 0-100
        - `force_notify`: Whether to force send, ignores alert threshold when true
        - `test_mode`: Test mode, only tests connection without sending actual report when true
      parameters:
      - description: WeChat notification trigger request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WeChatNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Triggered successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Invalid request parameters or WeChat notification not enabled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Manually trigger WeChat Enterprise notification
      tags:
      - Notification Management
  /notifications/wechat/status:
    get:
      consumes:
      - application/json
      description: Returns detailed information about WeChat Enterprise notification
        feature configuration status, connection status, recent notification history,
        etc.
      parameters:
      - default: 10
        description: History record count limit
        in: query
        maximum: 50
        minimum: 1
        name: history_limit
        type: integer
      - default: true
        description: Whether to include detailed configuration
        in: query
        name: include_config
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Status information retrieved successfully
          schema:
            $ref: '#/definitions/models.WeChatNotificationStatusResponse'
      summary: Get WeChat Enterprise notification status
      tags:
      - Notification Management
  /notifications/wechat/test:
    post:
      consumes:
      - application/json
      description: Send test message to WeChat Enterprise group to verify correctness
        and connectivity of Webhook URL configuration, will not send actual cost reports
      parameters:
      - description: Test request parameters (optional)
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.WeChatTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Connection test successful
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: WeChat notification not enabled or Webhook URL not configured
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Connection test failed or internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Test WeChat Enterprise Webhook connection
      tags:
      - Notification Management
...
schemes:
- http
- https
swagger: "2.0"